Move lib/Solver/SolverStats.h -> include/klee/SolverStats.h so that it is possible to ``#include "klee/util/ArrayExprHash.h"``
